refactor(kernel/memory): change constant type

+++ kernel/src/memory/mod.rs
@@ -4,7 +4,7 @@ use crate::{Output, init_end, init_start};
 
 mod physical;
 
-const PAGE_SIZE: usize = 0x1000;
+const PAGE_SIZE: u64 = 0x1000;
 
 pub fn init(entry: usize, descriptor_size: usize, descriptor_count: usize) {
     init_start!();
